refactor: move config.deps-installer to installing.env-installer
