+++ .changeset/refactor-workspace-domains.md
@@ -2,6 +2,7 @@
 "pnpm": minor
 "@pnpm/workspace.projects-filter": minor
 "@pnpm/workspace.projects-reader": minor
+"@pnpm/installing.env-installer": minor
 ---
 
-Refactor workspace domains: rename `project-finder` to `projects-reader`, merge `filter-packages-from-dir` into `filter-workspace-packages`, and rename it to `projects-filter`.
+Refactor workspace domains: rename `project-finder` to `projects-reader`, merge `filter-packages-from-dir` into `filter-workspace-packages`, and rename it to `projects-filter`. Also, move and rename `config/deps-installer` to `installing/env-installer`.
