+++ cli/utils/src/getConfig.ts
@@ -3,9 +3,9 @@ import path from 'node:path'
 
 import { formatWarn } from '@pnpm/cli.default-reporter'
 import { packageManager } from '@pnpm/cli.meta'
-import { resolveAndInstallConfigDeps } from '@pnpm/config.deps-installer'
 import { type CliOptions, type Config, getConfig as _getConfig } from '@pnpm/config.reader'
 import { requireHooks } from '@pnpm/hooks.pnpmfile'
+import { resolveAndInstallConfigDeps } from '@pnpm/installing.env-installer'
 import { createStoreController } from '@pnpm/store.connection-manager'
 import type { ConfigDependencies } from '@pnpm/types'
 import { lexCompare } from '@pnpm/util.lex-comparator'
