--- config/deps-installer/README.md
@@ -1,17 +1,0 @@
-# @pnpm/config.deps-installer
-
-> Installer for configurational dependencies
-
-<!--@shields('npm')-->
-[![npm version](https://img.shields.io/npm/v/@pnpm/config.deps-installer.svg)](https://www.npmjs.com/package/@pnpm/config.deps-installer)
-<!--/@-->
-
-## Installation
-
-```sh
-pnpm add @pnpm/config.deps-installer
-```
-
-## License
-
-MIT
