+++ engine/pm/commands/package.json
@@ -34,7 +34,6 @@
     "@pnpm/bins.linker": "workspace:*",
     "@pnpm/cli.meta": "workspace:*",
     "@pnpm/cli.utils": "workspace:*",
-    "@pnpm/config.deps-installer": "workspace:*",
     "@pnpm/config.reader": "workspace:*",
     "@pnpm/deps.graph-hasher": "workspace:*",
     "@pnpm/error": "workspace:*",
@@ -42,6 +41,7 @@
     "@pnpm/global.packages": "workspace:*",
     "@pnpm/installing.client": "workspace:*",
     "@pnpm/installing.deps-restorer": "workspace:*",
+    "@pnpm/installing.env-installer": "workspace:*",
     "@pnpm/lockfile.types": "workspace:*",
     "@pnpm/os.env.path-extender": "catalog:",
     "@pnpm/resolving.npm-resolver": "workspace:*",
