+++ installing/commands/package.json
@@ -39,7 +39,6 @@
     "@pnpm/cli.common-cli-options-help": "workspace:*",
     "@pnpm/cli.utils": "workspace:*",
     "@pnpm/colorize-semver-diff": "catalog:",
-    "@pnpm/config.deps-installer": "workspace:*",
     "@pnpm/config.matcher": "workspace:*",
     "@pnpm/config.pick-registry-for-package": "workspace:*",
     "@pnpm/config.reader": "workspace:*",
@@ -55,6 +54,7 @@
     "@pnpm/installing.context": "workspace:*",
     "@pnpm/installing.dedupe.check": "workspace:*",
     "@pnpm/installing.deps-installer": "workspace:*",
+    "@pnpm/installing.env-installer": "workspace:*",
     "@pnpm/lockfile.types": "workspace:*",
     "@pnpm/pkg-manifest.reader": "workspace:*",
     "@pnpm/pkg-manifest.utils": "workspace:*",
