+++ installing/env-installer/README.md
@@ -1,0 +1,17 @@
+# @pnpm/installing.env-installer
+
+> Installer for configurational dependencies
+
+<!--@shields('npm')-->
+[![npm version](https://img.shields.io/npm/v/@pnpm/installing.env-installer.svg)](https://www.npmjs.com/package/@pnpm/installing.env-installer)
+<!--/@-->
+
+## Installation
+
+```sh
+pnpm add @pnpm/installing.env-installer
+```
+
+## License
+
+MIT
