+++ installing/env-installer/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@pnpm/config.deps-installer",
+  "name": "@pnpm/installing.env-installer",
   "version": "1000.0.19",
   "description": "Installer for configurational dependencies",
   "keywords": [
@@ -9,8 +9,8 @@
   ],
   "license": "MIT",
   "funding": "https://opencollective.com/pnpm",
-  "repository": "https://github.com/pnpm/pnpm/tree/main/config/deps-installer",
-  "homepage": "https://github.com/pnpm/pnpm/tree/main/config/deps-installer#readme",
+  "repository": "https://github.com/pnpm/pnpm/tree/main/installing/env-installer",
+  "homepage": "https://github.com/pnpm/pnpm/tree/main/installing/env-installer#readme",
   "bugs": {
     "url": "https://github.com/pnpm/pnpm/issues"
   },
@@ -62,7 +62,7 @@
     "@pnpm/worker": "workspace:^"
   },
   "devDependencies": {
-    "@pnpm/config.deps-installer": "workspace:*",
+    "@pnpm/installing.env-installer": "workspace:*",
     "@pnpm/prepare": "workspace:*",
     "@pnpm/registry-mock": "catalog:",
     "@pnpm/testing.temp-store": "workspace:*",
