+++ installing/env-installer/src/index.ts
