+++ installing/env-installer/src/installConfigDeps.ts
