+++ installing/env-installer/src/migrateConfigDeps.ts
