+++ installing/env-installer/src/parseIntegrity.ts
