+++ installing/env-installer/src/pruneEnvLockfile.ts
