+++ installing/env-installer/src/resolveAndInstallConfigDeps.ts
