+++ installing/env-installer/src/resolveConfigDeps.ts
