+++ installing/env-installer/src/resolveManifestDependencies.ts
