+++ installing/env-installer/src/resolvePackageManagerIntegrities.ts
