+++ installing/env-installer/test/tsconfig.json
