+++ installing/env-installer/tsconfig.lint.json
